[REFACTOR]: model
