+++ src/main/java/id/ac/ui/cs/advprog/hiringgo/exception/GlobalExceptionHandler.java
@@ -1,19 +1,26 @@
 package id.ac.ui.cs.advprog.hiringgo.exception;
 
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.InvalidLogException;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.LogNotFoundException;
 import id.ac.ui.cs.advprog.hiringgo.model.WebResponse;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.FieldError;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.stream.Collectors;
 
-@ControllerAdvice
+@RestControllerAdvice
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(ConstraintViolationException.class)
@@ -62,6 +69,38 @@ public class GlobalExceptionHandler {
         WebResponse<?> response = WebResponse.builder()
                 .errors("Request body is missing or malformed")
                 .build();
+
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(LogNotFoundException.class)
+    public ResponseEntity<Map<String, Object>> handleLogNotFoundException(LogNotFoundException ex) {
+        Map<String, Object> response = new HashMap<>();
+        response.put("message", "Validasi log gagal");
+        response.put("errors", ex.getErrors());
+        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler(InvalidLogException.class)
+    public ResponseEntity<Map<String, Object>> handleInvalidLogException(InvalidLogException ex) {
+        Map<String, Object> response = new HashMap<>();
+        response.put("message", "Validasi log gagal");
+        response.put("errors", ex.getErrors());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<Map<String, Object>> handleValidationExceptions(MethodArgumentNotValidException ex) {
+        Map<String, String> errors = new HashMap<>();
+        ex.getBindingResult().getAllErrors().forEach(error -> {
+            String fieldName = ((FieldError) error).getField();
+            String errorMessage = error.getDefaultMessage();
+            errors.put(fieldName, errorMessage);
+        });
+
+        Map<String, Object> response = new HashMap<>();
+        response.put("message", "Validasi gagal");
+        response.put("errors", errors);
 
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
