--- src/main/java/id/ac/ui/cs/advprog/hiringgo/manajemenLog/exception/GlobalExceptionHandler.java
@@ -1,72 +1,0 @@
-package id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception;
-
-import jakarta.validation.ConstraintViolationException;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(LogNotFoundException.class)
-    public ResponseEntity<Map<String, Object>> handleLogNotFoundException(LogNotFoundException ex) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("message", "Validasi log gagal");
-        response.put("errors", ex.getErrors());
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
-    }
-    
-    @ExceptionHandler(InvalidLogException.class)
-    public ResponseEntity<Map<String, Object>> handleInvalidLogException(InvalidLogException ex) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("message", "Validasi log gagal");
-        response.put("errors", ex.getErrors());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, Object>> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getAllErrors().forEach(error -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        
-        Map<String, Object> response = new HashMap<>();
-        response.put("message", "Validasi gagal");
-        response.put("errors", errors);
-        
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<Map<String, Object>> handleConstraintViolation(ConstraintViolationException ex) {
-        Map<String, String> errors = new HashMap<>();
-        
-        ex.getConstraintViolations().forEach(violation -> {
-            String propertyPath = violation.getPropertyPath().toString();
-            String field = propertyPath.substring(propertyPath.lastIndexOf('.') + 1);
-            errors.put(field, violation.getMessage());
-        });
-        
-        Map<String, Object> response = new HashMap<>();
-        response.put("message", "Validasi gagal");
-        response.put("errors", errors);
-        
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Map<String, String>> handleGeneralException(Exception ex) {
-        Map<String, String> response = new HashMap<>();
-        response.put("error", "Terjadi kesalahan: " + ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
